some prep work
