+++ basculechecks/metricvalidator_test.go
@@ -401,44 +401,3 @@ func TestPrepMetrics(t *testing.T) {
 		})
 	}
 }
-
-func TestDeterminePartnerMetric(t *testing.T) {
-	tests := []struct {
-		description    string
-		partnersInput  []string
-		expectedResult string
-	}{
-		{
-			description:    "No Partners",
-			expectedResult: "none",
-		},
-		{
-			description:    "one wildcard",
-			partnersInput:  []string{"*"},
-			expectedResult: "wildcard",
-		},
-		{
-			description:    "one partner",
-			partnersInput:  []string{"TestPartner"},
-			expectedResult: "TestPartner",
-		},
-		{
-			description:    "many partners",
-			partnersInput:  []string{"partner1", "partner2", "partner3"},
-			expectedResult: "many",
-		},
-		{
-			description:    "many partners with wildcard",
-			partnersInput:  []string{"partner1", "partner2", "partner3", "*"},
-			expectedResult: "wildcard",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.description, func(t *testing.T) {
-			assert := assert.New(t)
-			partner := DeterminePartnerMetric(tc.partnersInput)
-			assert.Equal(tc.expectedResult, partner)
-		})
-	}
-}
